Next part of misspellings and other style issues fix

+++ Proximab/Proxima.Core/AI/AIResult.cs
@@ -46,7 +46,7 @@ namespace Proxima.Core.AI
         /// <summary>
         /// Gets the number of nodes per second (TotalNodes / Time).
         /// </summary>
-        public int NodesPerSecond => Time != 0 ? (int)(Stats.TotalNodes / Time) : 0;
+        public int NodesPerSecond => (int)Time != 0 ? (int)(Stats.TotalNodes / Time) : 0;
 
         /// <summary>
         /// Gets the number of nanoseconds per node ((Ticks / TotalNodes) * 100).
